select.go
select.go erstellt und get_zettel_list() aus main.go entfernt
